+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- allow configuring the program used to `open` associated files
+
 
 ## [2.1.0] - 2020-06-14